requirements.txt
streamlit
groq
python-dotenv
supabase
fuzzywuzzy
python-Levenshtein